// src/app/core/models/main/tours.model.ts
export class Tours{
  id: string;
  show: boolean;
  tenGV: string;
  email: string;
  sdt: string;
  noiDungThamQuan: string;
  khoa: string;
  nganh: string;
  soLuongSV: number;
  sinhVienNamThu: number;
  gvKhoa: string;
  soLuongGvThamGia: number;
  thoiGianDuKien: string;
  tenCongTy: string;
  diaChi: string;
  sdtLienHeCongTy: string;
  hocKi: string;
  namHoc: string;
  yeuCauCongVan: string;
  trangThai: boolean;
}